requirement.txt
# Core LangChain (stable with Python 3.10)
langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.16
langchain-google-genai==1.0.10
google-generativeai == 0.7.2
pinecone-client == 5.0.1

# Embeddings / NLP
sentence-transformers==2.6.1

transformers==4.39.3

# NumPy (CRITICAL FIX)
numpy==1.26.4

# App
flask==3.0.2
pypdf==4.2.0
python-dotenv==1.0.1

